Fix request isomorphic function, for now, the split is rather naive, still need improvement
