+++ source/Connection-node.js
@@ -1,0 +1,5 @@
+var utility = require('./utility/utility-browser');
+
+var Connection = module.exports = require('./Connection.js');
+
+Connection.prototype._request = utility.request;
