+++ source/Connection.js
@@ -1,5 +1,4 @@
-var utility = require('main').utility,
-    helpers = require('./utility/helpers.js'),
+var utility = require('./utility/utility.js'),
     ConnectionEvents = require('./connection/ConnectionEvents.js'),
     ConnectionStreams = require('./connection/ConnectionStreams.js'),
     ConnectionProfile = require('./connection/ConnectionProfile.js'),
@@ -45,7 +44,7 @@ function Connection() {
     this.username = settings.username;
     this.auth = settings.auth;
     if (settings.url) {
-      var urlInfo = helpers.urls.parseServerURL(settings.url);
+      var urlInfo = utility.urls.parseServerURL(settings.url);
       this.username = urlInfo.username;
       settings.hostname = urlInfo.hostname;
       settings.domain = urlInfo.domain;
@@ -64,7 +63,7 @@ function Connection() {
   }, settings);
 
   this.settings.domain = settings.domain ?
-      settings.domain : helpers.urls.defaultDomain;
+      settings.domain : utility.urls.defaultDomain;
 
   this.serverInfos = {
     // nowLocalTime - nowServerTime
@@ -291,7 +290,7 @@ Connection.prototype.request = function (params) {
     withoutCredentials = true;
   }
 
-  var request = utility.request({
+  var request = this._request({
     method : params.method,
     host : getHostname(this),
     port : this.settings.port,
@@ -436,14 +435,14 @@ Connection.login = function (params, callback) {
     'Content-Type': 'application/json'
   };
 
-  if (!helpers.isBrowser()) {
+  if (!utility.isBrowser()) {
     var origin = 'https://sw.';
     origin = params.origin ? origin + params.origin :
-    origin + helpers.urls.domains.client.production;
+    origin + utility.urls.domains.client.production;
     _.extend(headers, {Origin: origin});
   }
 
-  var domain = params.domain || helpers.urls.domains.client.production;
+  var domain = params.domain || utility.urls.domains.client.production;
 
   var pack = {
     method: 'POST',
@@ -475,7 +474,7 @@ Connection.login = function (params, callback) {
     }
   };
 
-  utility.request(pack);
+  this.request(pack);
 };
