+++ source/auth/Auth-browser.js
@@ -1,7 +1,6 @@
 /* global confirm, document, navigator, location, window */
 
 var utility = require('../utility/utility-browser.js'),
-  helpers = require('../utility/helpers.js'),
   Connection = require('../Connection.js'),
   _ = require('underscore');
 
@@ -393,7 +392,7 @@ Auth.prototype.internalError = function (message, jsonData) {
 Auth.prototype.login = function (settings) {
   this._checkCookies();
 
-  var defaultDomain = helpers.urls.defaultDomain;
+  var defaultDomain = utility.urls.defaultDomain;
   this.settings = settings = _.defaults(settings, {
     ssl: true,
     domain: defaultDomain
@@ -457,7 +456,7 @@ Auth.prototype.trustedLogout = function () {
  * @param settings: request settings
  */
 Auth.prototype.whoAmI = function (settings) {
-  var defaultDomain = helpers.urls.defaultDomain;
+  var defaultDomain = utility.urls.defaultDomain;
   this.settings = settings = _.defaults(settings, {
     ssl: true,
     domain: defaultDomain
@@ -512,7 +511,7 @@ Auth.prototype.whoAmI = function (settings) {
  * @returns {*}: a successful connection or false
  */
 Auth.prototype.loginWithCookie = function (settings) {
-  var defaultDomain = helpers.urls.defaultDomain;
+  var defaultDomain = utility.urls.defaultDomain;
   this.settings = settings = _.defaults(settings, {
     ssl: true,
     domain: defaultDomain
