+++ source/utility/utility.js
